edit dan update beberapa fitur baru

+++ routes/web.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Settings;
 use App\Http\Controllers\Post;
 use App\Http\Controllers\AdminController;
 use App\Http\Controllers\Posts;
+use App\Http\Controllers\Login;
 use App\Http\Controllers\Register;
 /*
 |--------------------------------------------------------------------------
@@ -19,29 +20,34 @@ use App\Http\Controllers\Register;
 |
 */
 
+//Homepage
 Route::get('/', [BlogController::class, 'index']);
 Route::get('/blog/{post:slug}', [BlogController::class, 'show']);
+
+
+
+//Login register auth
 Route::get('/register', [Register::class, 'index']);
-
-Route::get('/dashboard', [Dashboard::class, 'index'])->middleware('auth')->name('dashboard');
-Route::get('/dashboard/myposts',[BlogController::class, 'all']);
-Route::get('/dashboard/lists', [Dashboard::class, 'lists']);
-Route::get('/dashboard/settings', [Dashboard::class, 'setting']);
-Route::get('/dashboard/myaccount', [Dashboard::class, 'account']);
-Route::get('/dashboard/myposts/new', [Dashboard::class, 'create']);
-Route::post('/myposts/new/add', [Post::class, 'create']);
-
 Route::post('/register/add', [Register::class, 'register']);
-Route::get('/register/success', function (){
-  return view('succes');
-});
+Route::get('/register/success', [Register::class, 'suc']
+);
+Route::get('/login', [AdminController::class, 'index'])->middleware('guest')->name('login');
 
-Route::put('/dashboard/settings/update', [Dashboard::class, 'store']);
+Route::post('/login', [Login::class, 'auth']);
+Route::post('/logout', [Login::class, 'logout']);
 
-Route::put('/dashboard/myposts/{post:slug}/edit/save', [BlogController::class, 'update']);
-  Route::get('/dashboard/myposts/{post:slug}/edit', [BlogController::class, 'edit']);
-  Route::delete('/dashboard/myposts/{post:slug}/delete', [BlogController::class, 'destroy']);
 
-  Route::get('/login', [AdminController::class, 'index'])->middleware('guest')->name('login');
+//Dashboard 
+Route::get('/dashboard', [Dashboard::class, 'index'])->middleware('auth')->name('dashboard');
+Route::get('/dashboard/myposts',[BlogController::class, 'all'])->middleware('auth');
+Route::get('/dashboard/lists', [Dashboard::class, 'lists'])->middleware('auth');
+Route::get('/dashboard/settings', [Dashboard::class, 'setting'])->middleware('auth');
+Route::get('/dashboard/myaccount', [Dashboard::class, 'account'])->middleware('auth');
+Route::get('/dashboard/myposts/new', [Dashboard::class, 'create'])->middleware('auth');
+Route::post('/myposts/new/add', [Post::class, 'create'])->middleware('auth');
+Route::put('/dashboard/settings/update', [Dashboard::class, 'store'])->middleware('auth');
+Route::put('/dashboard/myposts/{post:slug}/edit/save', [BlogController::class, 'update'])->middleware('auth');
+  Route::get('/dashboard/myposts/{post:slug}/edit', [BlogController::class, 'edit'])->middleware('auth');
+  Route::delete('/dashboard/myposts/{post:slug}/delete', [BlogController::class, 'destroy'])->middleware('auth');
 
-Route::post('/login', [AdminController::class, 'auth']);
+
